Refactor mobile commands and MobileEventHandler
