+++ src/main/java/com/rettichlp/UnicacityAddon/commands/mobile/ASMSCommand.java
@@ -1,9 +1,10 @@
-package com.rettichlp.UnicacityAddon.commands;
+package com.rettichlp.UnicacityAddon.commands.mobile;
 
 import com.rettichlp.UnicacityAddon.base.abstraction.AbstractionLayer;
 import com.rettichlp.UnicacityAddon.base.abstraction.UPlayer;
 import com.rettichlp.UnicacityAddon.base.api.request.TabCompletionBuilder;
 import com.rettichlp.UnicacityAddon.base.registry.annotation.UCCommand;
+import com.rettichlp.UnicacityAddon.base.utils.TextUtils;
 import com.rettichlp.UnicacityAddon.events.MobileEventHandler;
 import net.minecraft.command.ICommand;
 import net.minecraft.command.ICommandSender;
@@ -22,7 +23,7 @@ import java.util.TimerTask;
  * @author RettichLP
  */
 @UCCommand
-public class ACallCommand implements IClientCommand {
+public class ASMSCommand implements IClientCommand {
 
     final Timer timer = new Timer();
     public static boolean isActive;
@@ -30,13 +31,13 @@ public class ACallCommand implements IClientCommand {
     @Override
     @Nonnull
     public String getName() {
-        return "acall";
+        return "asms";
     }
 
     @Override
     @Nonnull
     public String getUsage(@Nonnull ICommandSender sender) {
-        return "/acall [Spielername]";
+        return "/asms [Spielername] [Nachricht]";
     }
 
     @Override
@@ -53,7 +54,8 @@ public class ACallCommand implements IClientCommand {
     @Override
     public void execute(@Nonnull MinecraftServer server, @Nonnull ICommandSender sender, @Nonnull String[] args) {
         UPlayer p = AbstractionLayer.getPlayer();
-        if (args.length < 1) {
+
+        if (args.length < 2) {
             p.sendSyntaxMessage(getUsage(sender));
             return;
         }
@@ -70,7 +72,8 @@ public class ACallCommand implements IClientCommand {
                     return;
                 }
 
-                p.sendChatMessage("/call " + number);
+                String message = TextUtils.makeStringByArgs(args, " ").replace(args[0], "");
+                p.sendChatMessage("/sms " + number + message);
             }
         }, 250L);
     }
